requirements.txt
transformers
datasets
trl
peft
python-dotenv